requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py>=0.8.1            # via tensorboard, tensorflow
amqp>=2.5.2               # via kombu
appnope>=0.1.0            # via ipython
asgiref>=3.2.3            # via django
astor>=0.8.0              # via tensorflow
backcall>=0.1.0           # via ipython
billiard>=3.6.1.0         # via celery
brotlipy>=0.7.0
cachetools>=3.1.1         # via google-auth
celery[redis]>=4.3.0
certifi>=2019.11.28       # via requests, sentry-sdk
cffi>=1.13.2              # via brotlipy, cryptography
chardet>=3.0.4            # via requests
cryptography>=2.8         # via django-rest-knox
dataclasses>=0.6          # via djongo
decorator>=4.4.1          # via ipython, traitlets
defusedxml>=0.6.0         # via odfpy
diff-match-patch>=20181111  # via django-import-export
dj-database-url>=0.5.0
django-import-export>=2.0
django-js-reverse>=0.9.1
django-log-request-id>=1.4.0
django-model-utils>=3.2.0
django-rest-knox>=4.1.0
django-webpack-loader>=0.6.0
django==2.2.7
djangorestframework>=3.10.3
djongo>=1.2.38
dnspython>=1.16.0
et-xmlfile>=1.0.1         # via openpyxl
gast>=0.2.2               # via tensorflow
geographiclib>=1.50       # via geopy
geopy>=1.20.0
google-auth-oauthlib>=0.4.1  # via tensorboard
google-auth>=1.8.1        # via google-auth-oauthlib, tensorboard
google-pasta>=0.1.8       # via tensorflow
grpcio>=1.25.0            # via tensorboard, tensorflow
gunicorn>=20.0.4
h5py>=2.10.0              # via keras-applications
idna>=2.8                 # via requests
importlib-metadata>=1.2.0  # via kombu
ipython-genutils>=0.2.0   # via traitlets
ipython>=7.10.1
jdcal>=1.4.1              # via openpyxl
jedi>=0.15.1              # via ipython
joblib>=0.14.0            # via scikit-learn
keras-applications>=1.0.8  # via tensorflow
keras-preprocessing>=1.1.0  # via tensorflow
kombu>=4.6.7              # via celery
markdown>=3.1.1           # via tensorboard
markuppy>=1.14            # via tablib
mongoengine>=0.18.2
more-itertools>=8.0.2     # via zipp
numpy>=1.17.4             # via h5py, keras-applications, keras-preprocessing, opt-einsum, pandas, scikit-learn, scipy, tensorboard, tensorflow, xgboost
oauthlib>=3.1.0           # via requests-oauthlib
odfpy>=1.4.0              # via tablib
openpyxl>=3.0.2           # via tablib
opt-einsum>=3.1.0         # via tensorflow
pandas>=0.25.3
parso>=0.5.1              # via jedi
pexpect>=4.7.0            # via ipython
pickleshare>=0.7.5        # via ipython
prompt-toolkit>=3.0.2     # via ipython
protobuf>=3.11.1          # via tensorboard, tensorflow
psutil>=5.6.7
ptyprocess>=0.6.0         # via pexpect
pyasn1-modules>=0.2.7     # via google-auth
pyasn1>=0.4.8             # via pyasn1-modules, rsa
pycparser>=2.19           # via cffi
pygments>=2.5.2           # via ipython
pymongo>=3.9.0
python-dateutil>=2.8.1    # via pandas
python-decouple>=3.3
pytz>=2019.3              # via celery, django, pandas
pyyaml>=5.2               # via tablib
redis>=3.3.11             # via celery
requests-oauthlib>=1.3.0  # via google-auth-oauthlib
requests>=2.22.0          # via requests-oauthlib, tensorboard
rsa>=4.0                  # via google-auth
scikit-learn>=0.22        # via sklearn
scipy>=1.3.3              # via scikit-learn, xgboost
sentry-sdk>=0.13.5
six>=1.13.0               # via absl-py, cryptography, google-auth, google-pasta, grpcio, h5py, keras-preprocessing, mongoengine, protobuf, python-dateutil, tensorboard, tensorflow, traitlets
sklearn>=0.0
sqlparse>=0.2.4           # via django, djongo
tablib>=0.14.0            # via django-import-export
tensorboard>=2.0.2        # via tensorflow
# tensorflow-estimator>=2.0.1  # via tensorflow
# tensorflow>=2.0.0
termcolor>=1.1.0          # via tensorflow
traitlets>=4.3.3          # via ipython
urllib3>=1.25.7           # via requests, sentry-sdk
vine>=1.3.0               # via amqp, celery
wcwidth>=0.1.7            # via prompt-toolkit
werkzeug>=0.16.0          # via tensorboard
wheel>=0.33.6             # via tensorboard, tensorflow
whitenoise>=4.1.4
wrapt>=1.11.2             # via tensorflow
xgboost>=0.90
xlrd>=1.2.0               # via tablib
xlwt>=1.3.0               # via tablib
zipp>=0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
